requirements.txt
python-telegram-bot
google-generativeai
Pillow
python-dotenv
langdetect
peewee
psycopg2-binary
